+++ NEWS.md
@@ -1,3 +1,7 @@
+# Version 1.2.0
+ - Time-to-quilibrium (tteq) function
+ - Req function
+
 # Version 1.1.0
  - Added baseline drift parameter
  - Added global offset parameter